+++ Frends.AzureDataLake.DeleteDirectory/Frends.AzureDataLake.DeleteDirectory.Tests/UnitTests.cs
@@ -5,7 +5,6 @@ using System.Globalization;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
-using Azure;
 using Azure.Storage.Files.DataLake;
 using dotenv.net;
 using Frends.AzureDataLake.DeleteDirectory.Definitions;
@@ -70,7 +69,7 @@ public class UnitTests
         await directoryClient.CreateIfNotExistsAsync(null, CancellationToken.None);
 
         var deleted = await AzureDataLake.DeleteDirectory(
-            input, CancellationToken.None
+            input, new Options(), CancellationToken.None
         );
 
         Assert.IsTrue(deleted.DirectoryWasDeleted);
@@ -79,7 +78,7 @@ public class UnitTests
     }
 
     [TestMethod]
-    [ExpectedException(typeof(RequestFailedException))]
+    [ExpectedException(typeof(Exception))]
     public async Task TestDeleteDirectory_throws_ContainerNotFound()
     {
         var input = new Input
@@ -89,31 +88,32 @@ public class UnitTests
             DirectoryName = _directoryName
         };
 
-        await AzureDataLake.DeleteDirectory(input, CancellationToken.None);
+        await AzureDataLake.DeleteDirectory(input, new Options(), CancellationToken.None);
     }
 
     [TestMethod]
-    [ExpectedException(typeof(ArgumentNullException))]
+    [ExpectedException(typeof(Exception))]
     public async Task TestDeleteDirectory_ThrowsParameterEmpty_WhenContainerNameIsNull()
     {
         await AzureDataLake.DeleteDirectory(
             new Input { ConnectionString = _connectionString, ContainerName = null, DirectoryName = _directoryName },
+            new Options(),
             CancellationToken.None
         );
     }
 
     [TestMethod]
-    [ExpectedException(typeof(FormatException))]
+    [ExpectedException(typeof(Exception))]
     public async Task TestDeleteDirectory_throws_ParameterNotValid()
     {
-        await AzureDataLake.DeleteDirectory(new Input { ConnectionString = "Not valid parameter", ContainerName = "Valid name", DirectoryName = "Valid name" }, CancellationToken.None);
+        await AzureDataLake.DeleteDirectory(new Input { ConnectionString = "Not valid parameter", ContainerName = "Valid name", DirectoryName = "Valid name" }, new Options(), CancellationToken.None);
     }
 
     [TestMethod]
-    [ExpectedException(typeof(FormatException))]
+    [ExpectedException(typeof(Exception))]
     public async Task TestDeleteDirectory_throws_ClientNotFound()
     {
-        await AzureDataLake.DeleteDirectory(new Input { ConnectionString = "DefaultEndpointsProtocol=https;AccountName=unitTestStorage;AccountKey=abcdefghijklmnopqrstuyxz123456789;EndpointSuffix=core.windows.net", ContainerName = _containerName, DirectoryName = _directoryName }, CancellationToken.None);
+        await AzureDataLake.DeleteDirectory(new Input { ConnectionString = "DefaultEndpointsProtocol=https;AccountName=unitTestStorage;AccountKey=abcdefghijklmnopqrstuyxz123456789;EndpointSuffix=core.windows.net", ContainerName = _containerName, DirectoryName = _directoryName }, new Options(), CancellationToken.None);
     }
 
     [TestMethod]
@@ -139,7 +139,7 @@ public class UnitTests
         var directoryClient = AzureDataLake.GetDataLakeDirectory(input);
         await directoryClient.CreateIfNotExistsAsync(null, CancellationToken.None);
 
-        var result = await AzureDataLake.DeleteDirectory(input, default);
+        var result = await AzureDataLake.DeleteDirectory(input, new Options(), default);
         Assert.IsTrue(result.DirectoryWasDeleted);
         var directoryExists = await directoryClient.ExistsAsync(CancellationToken.None);
         Assert.IsFalse(directoryExists);